+++ packages/core-js/actual/array/index.js
@@ -3,6 +3,7 @@ require('../../modules/es.map');
 require('../../modules/es.object.to-string');
 require('../../modules/esnext.array.find-last');
 require('../../modules/esnext.array.find-last-index');
+require('../../modules/esnext.array.from-async');
 require('../../modules/esnext.array.group');
 require('../../modules/esnext.array.group-by');
 require('../../modules/esnext.array.group-by-to-map');