+++ packages/core-js/full/array/index.js
@@ -1,6 +1,5 @@
 var parent = require('../../actual/array');
 require('../../modules/es.promise');
-require('../../modules/esnext.array.from-async');
 // TODO: Remove from `core-js@4`
 require('../../modules/esnext.array.at');
 // TODO: Remove from `core-js@4`